reorganize home page and move out components
